Restructure file system according to RubyGems guides
