+++ README.md
@@ -1,20 +1,20 @@
-# ActiveadminAsyncExporter
+# ActiveAdmin::AsyncExporter
 
 Async exporter for Active Admin using Sidekiq.
 
 ## Installation
 All you have to do is add the gem to your project like this:
 ```bash
-$ gem 'activeadmin_async_exporter'
+$ gem 'activeadmin-async_exporter'
 ```
 
 ## Contributing
-Bug reports (please use Issues) and pull requests are welcome on GitHub at https://github.com/rootstrap/activeadmin_async_exporter. This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct.
+Bug reports (please use Issues) and pull requests are welcome on GitHub at https://github.com/rootstrap/activeadmin-async_exporter. This project is intended to be a safe, welcoming space for collaboration, and contributors are expected to adhere to the [Contributor Covenant](http://contributor-covenant.org) code of conduct.
 
 ## License
 The gem is available as open source under the terms of the [MIT License](https://opensource.org/licenses/MIT).
 
 ## Credits
-**ActiveadminAsyncExporter** is maintained by [Rootstrap](http://www.rootstrap.com) with the help of our [contributors](https://github.com/rootstrap/activeadmin_async_exporter/contributors).
+**ActiveAdmin::AsyncExporter** is maintained by [Rootstrap](http://www.rootstrap.com) with the help of our [contributors](https://github.com/rootstrap/activeadmin-async_exporter/contributors).
 
 [<img src="https://s3-us-west-1.amazonaws.com/rootstrap.com/img/rs.png" width="100"/>](http://www.rootstrap.com)
