+++ lib/active_admin/async_exporter/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module ActiveAdmin
+  module AsyncExporter
+    VERSION = '0.1.0'
+  end
+end
