+++ lib/activeadmin-async_exporter.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require 'active_admin/async_exporter'
