--- lib/activeadmin_async_exporter/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module ActiveadminAsyncExporter
-  VERSION = '0.1.0'
-end
