+++ lib/generators/active_admin/async_exporter/install/templates/admin_report.rb.tt
