+++ lib/generators/active_admin/async_exporter/install/templates/admin_reports.rb.tt
