+++ lib/generators/active_admin/async_exporter/install/templates/migration.rb.tt
